wip internal refactoring
