+++ scoringrules/core/crps/__init__.py
@@ -1,3 +1,3 @@
 from .ensemble import ensemble
 from .closed import normal, lognormal, logistic
-from .guguncs import
+from .guguncs import estimator_gufuncs
